Forms_images/requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
openai>=1.12.0
Pillow>=10.2.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0